feat: reusable ProcessesTab with per-app process views, fix clear failed runs
Extract ProcessesTab component from DevTools for reuse in app detail views.
Each app now has a Processes tab showing its PM2 processes with live logs
and controls. Fix deleteFailedRuns sending wrong query parameter.
