+++ .changelog/v1.25.x.md
@@ -1,0 +1,20 @@
+# Release v1.25.x - Changelog
+
+Released: YYYY-MM-DD
+
+## Overview
+
+Reusable ProcessesTab component with per-app process views and a bug fix for clearing failed AI runs.
+
+## 🎉 Features
+
+- **App Processes Tab**: Each app's detail view now has a "Processes" tab showing only that app's PM2 processes with live logs, restart/stop/start controls
+- **Reusable ProcessesTab**: Extracted process management UI into a shared component used by both the DevTools Processes page and app detail views
+
+## 🐛 Fixes
+
+- **Clear Failed Runs**: Fixed "Only filter=failed is supported" error when clicking Clear Failed on the DevTools Runs page — client was sending wrong query parameter
+
+## 🔧 Improvements
+
+- **Alphabetical app tabs**: App detail tabs are now alphabetically ordered (Automation, Documents, Git, GSD, Processes, Tasks)
