+++ client/src/components/apps/constants.js
@@ -1,8 +1,9 @@
 export const APP_DETAIL_TABS = [
   { id: 'overview', label: 'Overview' },
-  { id: 'tasks', label: 'Tasks' },
   { id: 'automation', label: 'Automation' },
   { id: 'documents', label: 'Documents' },
   { id: 'git', label: 'Git' },
-  { id: 'gsd', label: 'GSD' }
+  { id: 'gsd', label: 'GSD' },
+  { id: 'processes', label: 'Processes' },
+  { id: 'tasks', label: 'Tasks' },
 ];
