+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portos",
-  "version": "1.24.1",
+  "version": "1.25.0",
   "private": true,
   "description": "Local dev machine App OS portal",
   "author": "Adam Eivy (@antic|@atomantic)",
